requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2018.03.07=0
certifi=2018.4.16=py36_0
cffi=1.11.5=py36h9745a5d_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.1.2=cuda9.0_0
intel-openmp=2018.0.3=0
libedit=3.1=heed3624_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=hdf63c60_3
libgfortran-ng=7.2.0=hdf63c60_3
libstdcxx-ng=7.2.0=hdf63c60_3
mkl=2018.0.3=1
mkl_fft=1.0.2=py36h651fb7a_0
mkl_random=1.0.1=py36h4414c95_1
nccl=1.3.5=cuda9.0_0
ncurses=6.0=h9df7e31_2
ninja=1.8.2=py36h6bb024c_1
numpy=1.14.5=py36h1b885b7_4
numpy-base=1.14.5=py36hdbf6ddf_4
openssl=1.0.2o=h20670df_0
pip=9.0.3=py36_0
pycparser=2.18=py36_1
python=3.6.5=hc3d631a_0
pytorch=0.4.0=py36hdf912b8_0
readline=7.0=ha6073c6_4
setuptools=39.0.1=py36_0
sqlite=3.23.1=he433501_0
tk=8.6.7=hc745277_3
wheel=0.31.0=py36_0
xz=5.2.3=h55aa19d_2
zlib=1.2.11=ha838bed_2
